feat: create types library for easier sharing of enums
