+++ libs/wallet-types/README.md
@@ -1,0 +1,11 @@
+# wallet-types
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build wallet-types` to build the library.
+
+## Running unit tests
+
+Run `nx test wallet-types` to execute the unit tests via [Jest](https://jestjs.io).
