+++ libs/wallet-types/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@vegaprotocol/wallet-types",
+  "version": "0.0.1",
+  "type": "commonjs"
+}
