+++ libs/wallet-ui/src/components/transaction-history/transaction-item.tsx
@@ -2,7 +2,7 @@ import { useExplorerUrl } from '../../hooks/use-explorer-url'
 import { formatDate } from '../../lib/date'
 import type { Transaction } from '../../lib/transactions'
 import { truncateMiddle } from '../../lib/truncate-middle'
-import { TRANSACTION_TITLES } from '../../lib/transactions'
+import { TRANSACTION_TITLES } from '@vegaprotocol/wallet-types'
 import { ButtonUnstyled } from '../button-unstyled'
 import { ExternalLink } from '../external-link'
 import { ArrowTopRight } from '../icons/arrow-top-right'
